Other Org Aggregator Performance Refactor
Addressing some of the issues with the Other Org Aggregator.

Queries performed on each project have been aggreagated to be run
against all other-org projects in a single go.  This appears to have
reduced the time taken to aggreate other orgs from around 17 seconds to
10 seconds for the sample size on my machine.  This needs further smoke
testing to ensure that reformated queries are achieveing the same
results